+++ objects/core.cpp
@@ -14,7 +14,7 @@ Value next_core_clock(const Value *args, int numargs) {
 	// TODO: clock_t is 64bits long, casting it to 32bits will reset the time
 	// every 36 mins or so. Casting it to double will probably result losing
 	// some precision.
-	return Value((double)clock());
+	return Value(clock());
 }
 
 Value next_core_type_of(const Value *args, int numargs) {
@@ -74,7 +74,7 @@ void addClocksPerSec() {
 	GcObject::CoreContext->add_public_mem(n);
 	int s = GcObject::CoreContext->get_mem_slot(n);
 	GcObject::CoreContext->defaultConstructor->bcc->push(
-	    Value((double)CLOCKS_PER_SEC));
+	    Value(CLOCKS_PER_SEC));
 	GcObject::CoreContext->defaultConstructor->bcc->store_object_slot(s);
 }
